Move orders to Vehicles folder
