+++ src/OpenLoco/Windows/Map.cpp
@@ -11,7 +11,6 @@
 #include "../Objects/ObjectManager.h"
 #include "../Objects/RoadObject.h"
 #include "../Objects/TrackObject.h"
-#include "../Orders.h"
 #include "../StationManager.h"
 #include "../Things/Thing.h"
 #include "../Things/ThingManager.h"
@@ -19,6 +18,7 @@
 #include "../Types.hpp"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
+#include "../Vehicles/Orders.h"
 #include "../Vehicles/Vehicle.h"
 #include "../Widget.h"
