+++ src/OpenLoco/Windows/Vehicle.cpp
@@ -15,13 +15,13 @@
 #include "../Objects/TrackObject.h"
 #include "../Objects/WaterObject.h"
 #include "../OpenLoco.h"
-#include "../Orders.h"
 #include "../StationManager.h"
 #include "../Things/ThingManager.h"
 #include "../TrackData.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
+#include "../Vehicles/Orders.h"
 #include "../ViewportManager.h"
 #include "../Widget.h"
 #include <map>
