+++ src/OpenLoco/openloco.vcxproj
@@ -80,7 +80,6 @@
     <ClCompile Include="Objects\WallObject.cpp" />
     <ClCompile Include="Objects\WaterObject.cpp" />
     <ClCompile Include="OpenLoco.cpp" />
-    <ClCompile Include="Orders.cpp" />
     <ClCompile Include="Paint\Paint.cpp" />
     <ClCompile Include="Paint\PaintEntity.cpp" />
     <ClCompile Include="Paint\PaintVehicle.cpp" />
@@ -109,6 +108,7 @@
     <ClCompile Include="Utility\Numeric.cpp" />
     <ClCompile Include="Utility\String.cpp" />
     <ClCompile Include="Vehicles\CreateVehicle.cpp" />
+    <ClCompile Include="Vehicles\Orders.cpp" />
     <ClCompile Include="Vehicles\Vehicle.cpp" />
     <ClCompile Include="Vehicles\VehicleBody.cpp" />
     <ClCompile Include="Vehicles\VehicleHead.cpp" />
@@ -254,7 +254,6 @@
     <ClInclude Include="Objects\WallObject.h" />
     <ClInclude Include="Objects\WaterObject.h" />
     <ClInclude Include="OpenLoco.h" />
-    <ClInclude Include="Orders.h" />
     <ClInclude Include="Paint\Paint.h" />
     <ClInclude Include="Paint\PaintEntity.h" />
     <ClInclude Include="Paint\PaintVehicle.h" />
@@ -287,6 +286,7 @@
     <ClInclude Include="Utility\Stream.hpp" />
     <ClInclude Include="Utility\String.hpp" />
     <ClInclude Include="Utility\Yaml.hpp" />
+    <ClInclude Include="Vehicles\Orders.h" />
     <ClInclude Include="Vehicles\Vehicle.h" />
     <ClInclude Include="Viewport.hpp" />
     <ClInclude Include="ViewportManager.h" />
